+++ src/functions/manageConfig.js
@@ -12,12 +12,13 @@ const Checks = require("../classes/Checks");
  * The `whatShouldIDo` parameter defines what this
  * function should do. The available values are:
  * - `'create'` - Creates the configuration file.
- * - `'delete'` - Delete the entire configuration file.
+ * - `'delete'` - Deletes the entire configuration file.
+ * - `'remove'` - Removes a specific key from the configuration file.
  * - `'get'` - Get the contents of the configuration file in both raw and parsed format.
  * - `'add'` - Add a value to the configuration file. Appends to an existing array if the parameter already exists.
  *
  * @param {string} whatShouldIDo One of the above values.
- * @param {object} data An object structure that will be converted to JSON. Only needed if `whatShouldIDo` is `'add'`.
+ * @param {object | string} data Either an object structure that will be converted to JSON, or the key to delete.
  * @returns Either data, `true` if the operation was successful, or `false` if it was unsuccessful.
  */
 const _manageConfig = (whatShouldIDo, data) => {
@@ -45,6 +46,16 @@ const _manageConfig = (whatShouldIDo, data) => {
     } catch (_) {
       return false;
     }
+  } else if (whatShouldIDo === "remove") {
+    try {
+      const contents = JSON.parse(fs.readFileSync(configPath, { flag: "r", encoding: "utf8" }));
+      delete contents[data];
+
+      fs.writeFileSync(configPath, JSON.stringify(contents));
+      return true;
+    } catch (_) {
+      return false;
+    }
   } else if (whatShouldIDo === "get") {
     try {
       const contents = fs.readFileSync(configPath, { flag: "r", encoding: "utf8" });
@@ -60,9 +71,14 @@ const _manageConfig = (whatShouldIDo, data) => {
       const contents = JSON.parse(fs.readFileSync(configPath, { flag: "r", encoding: "utf8" }));
 
       for (const param in data) {
-        if (Array.isArray(contents[param])) {
+        if (param === "history" && Array.isArray(data[param])) {
+          // Replace the history array entirely
+          contents[param] = data[param];
+        } else if (Array.isArray(contents[param])) {
+          // Append to other arrays if they exist
           contents[param] = contents[param].concat(data[param]);
         } else {
+          // Add new data
           contents[param] = data[param];
         }
       }